Add personal message to index and style files
As the idea of this lesson file is to simply add some
personal alterations, I added a div with an id of
"kestrel-special" and wrote a meassage/signature
using a nested paragraph element inside the div.
I then restyled my kestrel-special div using the style
sheet, including a hover effect just for fun.